Add Ability to Delete a DNS Record
A DNS record can be deleted by specifying the -Z option. A DNS record is
only deleted if it matches the type (-t TYPE), name (-n NAME) and
content (-c CONTENT) passed to the script. As a sanity check, the user
is prompted to confirm deletion of a matching DNS record. This prompt
can be suppressed using the -a option.