Include Both API Token and API Legacy Key in Request
Previously, either the Cloudflare API token or the legacy API key credentials
were included in the request depending on the user's choice. This may have
generated a malformed request which Curl appeared to deal with.

Including both he Cloudflare API token or the legacy API key credentials
ensures the request is not malformed.

If an API token is used, empty X-Auth-Email and X-Auth-Key headers are sent.
If a legacy API key is used, an empty Authorization header is sent. In either
case, the empty headers are ignored by Curl and have no adverse affect.